+++ client/web/.env.development
@@ -1,2 +1,1 @@
-SERVER_TRPC_DOMAIN=localhost:8787
 NEXT_PUBLIC_SERVER_API_HOST=http://localhost:8787
