+++ client/web/src/app/page.tsx
@@ -1,4 +1,3 @@
-import { TrpcTest } from "@/test/components/trpc-test";
 import { React95Test } from "@/test/components/react95-test";
 
 import { titleCss } from "./page.css";
@@ -9,7 +8,6 @@ export default function Home() {
     <main>
       <h1 className={titleCss}>Hello GitCy World!</h1>
       <HonoTest />
-      <TrpcTest />
       <React95Test />
     </main>
   );
