+++ client/web/src/common/api/query-provider.tsx
@@ -1,0 +1,16 @@
+"use client";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+import { ReactNode, useState } from "react";
+
+interface Props {
+  children: ReactNode;
+}
+
+export const QueryProvider = ({ children }: Props) => {
+  const [queryClient] = useState(() => new QueryClient());
+
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+};
