--- client/web/src/common/trpc/index.tsx
@@ -1,5 +1,0 @@
-"use client";
-import { createTRPCReact } from "@trpc/react-query";
-import type { AppRouter } from "@repo/server-app";
-
-export const trpc = createTRPCReact<AppRouter>();
