--- client/web/src/common/trpc/provider.tsx
@@ -1,32 +1,0 @@
-"use client";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { httpBatchLink } from "@trpc/client";
-import { ReactNode, useState } from "react";
-import { trpc } from ".";
-
-interface Props {
-  children: ReactNode;
-}
-
-export const QueryProvider = ({ children }: Props) => {
-  const [queryClient] = useState(() => new QueryClient());
-  const [trpcClient] = useState(() =>
-    trpc.createClient({
-      links: [
-        httpBatchLink({
-          url: "/trpc",
-          // You can pass any HTTP headers you wish here
-          async headers() {
-            return {};
-          },
-        }),
-      ],
-    })
-  );
-
-  return (
-    <trpc.Provider client={trpcClient} queryClient={queryClient}>
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </trpc.Provider>
-  );
-};
