+++ server/app/src/index.ts
@@ -1,2 +1,1 @@
-export type { AppRouter } from "./app";
 export type { ApiRouter } from "./router/api";
